requirements.txt
numpy
pandas
scipy
scikit-learn
xgboost
rasterio
matplotlib
PyYAML
joblib
pytest
